Create Snipe-it-Agent.ps1
This is a Powershell Script for Snipe-it to create new Asset&Model or Update yout Asset specs in Snipe-it.

Run it on the device whose data you want to upload to your Snipe-it Server. It is important that the device is in a network in which it reaches the server.